example/requirement.txt
kafka-python==2.0.2
confluent-kafka==2.3.0
requests
beautifulsoup4

